Adapter: refactor protocols and importers
