+++ adapter/src/adapter/importer/index.ts
@@ -1,0 +1,7 @@
+import { HttpImporter } from './HttpImporter';
+
+export * from './Importer';
+
+export const Protocol = {
+  HTTP: new HttpImporter(),
+};
